Compile boot rom with regular sh2eb-linux-musl toolchain.
Previously, we had to download and use a separate bare metal toolchain to
compile the boot rom sources, due to symbol prefix inconsistency.

This change represents the minimal subset of prefix modifications required to
compile the boot rom sources with the regular sh2eb-linux-musl-gcc toolchain
(or any other normal toolchain).

Note, that there are still a number of hard-coded underscore-prefixed variable
names in assembly source and in inline assembly in c files. A separate patch
will address the remaining underscore-prefixed symbol names.